Moved command/utils.py to utils directory
